Tests/LibWeb/Text/expected/parser-recursion-depth-limit.txt

unary-deep: SyntaxError
parens-deep: SyntaxError
array-deep: SyntaxError
block-deep: SyntaxError
binding-deep: SyntaxError
shallow: parsed
